rx: Refactor rxi_ReceivePacket call selection
Refactor the call selection logic in rxi_ReceivePacket so that it is
a little bit easier to follow, and better optimised to the common case.

Split the current logic into a function for packets being received by
a server, and a function for packets being received by a client.

Reviewed-on: http://gerrit.openafs.org/8297
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Derrick Brashear <shadow@your-file-system.com>
(cherry picked from commit 3c6570413c94afe4107634e9ca3f923f3cd30c21)

Change-Id: I0ac84fe47398b92a24e6c34cd4f0036e7bb27fcb
Reviewed-on: http://gerrit.openafs.org/8394
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>